Add Honeycomb marker before starting the deploy
The rollout of the new version has already started by the time our
deploy step has finished. In the case where the deploy action takes
longer than expected for some reason, adding the deploy marker
afterwards means we've got some production impact happening without a
marker for a little while.

The downside to swapping it round is that we'll have rogue markers if
the deploy doesn't happen for some reason. We can probably keep a
reference to the marker and update it to indicate it failed to solve
that.