Push services images to AWS ECR as well
I've manually pushed some services images to AWS ECR to get things
going. This commit is a placeholder to show how we might automate this
in future.

For this work to be completed, we need to set up and IAM role in the
AWS account which will trust GitHub's OIDC provider, and which has
push access to ECR.

There are docs here:
https://docs.github.com/en/actions/deployment/security-hardening-your-deployments/configuring-openid-connect-in-amazon-web-services