Set org.opencontainers.image.title explicitly
I had an action fail when the built image had a name that differed from the
repository name:
https://github.com/replicate/infra/actions/runs/8817778926/job/24205953769

docker/metadata-action assumes that org.opencontainers.image.title should be the
repoository name, and we use this label for the `image` output, which is wrong.

We can fix this by overriding the default value of the
org.opencontainers.image.title label to the `image` input.