Pass through `file` arg to docker/build-push-action
in a way that doesn't blow up. This is based on the `v5` action not
using a default value for `file` and then the javascript truthiness of
the value being used to decide to append a `--file` argument.

An empty string is false-y and `||` is allowed within a github actions
expression, so this should be `undefined || ''` when no `file` value is
provided, meaning no `--file` argument is appended to the build
arguments.